machine_learning_pipeline/requirements.txt
# Build dependencies (must be first)
setuptools>=65.0.0
wheel>=0.37.0
Cython>=0.29.0

# Core ML and Data Processing
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
pyspark>=3.4.0,<4.0.0

# Visualization
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0
plotly>=5.15.0,<6.0.0

# Configuration and Utilities
PyYAML>=6.0.0,<7.0.0
click>=8.1.0,<9.0.0

# Monitoring
evidently==0.4.25
scipy>=1.11.0,<2.0.0

# Jupyter and Development
jupyterlab>=4.0.0,<5.0.0
ipykernel>=6.25.0,<7.0.0

# Airflow 
apache-airflow>=2.7.0,<3.0.0